fix: some reward bug
